# docs/swagger.yaml
definitions:
  cmd_api-gateway.apiKeyCreateReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  cmd_api-gateway.apiKeyCreateResp:
    properties:
      created_at:
        type: string
      id:
        type: string
      key:
        type: string
      name:
        type: string
      scopes:
        items:
          type: string
        type: array
    type: object
  cmd_api-gateway.apiKeysResp:
    properties:
      keys:
        items:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_apikey.APIKey'
        type: array
    type: object
  cmd_api-gateway.artifactReq:
    properties:
      content:
        type: string
      conversation_id:
        type: string
      language:
        type: string
      name:
        type: string
      type:
        type: string
    required:
    - content
    - conversation_id
    - name
    - type
    type: object
...
  cmd_api-gateway.conversationsResp:
    properties:
      conversations:
        items:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_agent_model.Conversation'
        type: array
    type: object
  cmd_api-gateway.filesResp:
    properties:
      files:
        items:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_fileupload.Upload'
        type: array
    type: object
...
  cmd_api-gateway.loginResp:
    properties:
      token:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.TokenPair'
      user:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.User'
    type: object
  cmd_api-gateway.logoutResp:
    properties:
      message:
        type: string
    type: object
  cmd_api-gateway.meResp:
    properties:
      user:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.User'
    type: object
...
  cmd_api-gateway.refreshResp:
    properties:
      token:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.TokenPair'
      user:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.User'
    type: object
  cmd_api-gateway.registerResp:
    properties:
      user:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.User'
    type: object
...
  github_com_r0lm0_go-saas-api_internal_agent_model.Artifact:
    properties:
      content:
        type: string
      conversation_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_deleted:
        type: boolean
      language:
        type: string
      message_id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  github_com_r0lm0_go-saas-api_internal_agent_model.Conversation:
    properties:
      agent_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      messages:
        items:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_agent_model.Message'
        type: array
      metadata:
        additionalProperties: {}
        type: object
      status:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_agent_model.ConversationStatus'
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_r0lm0_go-saas-api_internal_agent_model.ConversationStatus:
    enum:
    - active
    - archived
    - deleted
    type: string
    x-enum-varnames:
    - ConversationActive
    - ConversationArchived
    - ConversationDeleted
  github_com_r0lm0_go-saas-api_internal_agent_model.Message:
    properties:
      content:
        type: string
      conversation_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      latency_ms:
        type: integer
      model:
        type: string
      role:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_agent_model.MessageRole'
      tenant_id:
        type: string
      tokens_input:
        type: integer
      tokens_output:
        type: integer
      tool_calls:
        items:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_agent_model.ToolCall'
        type: array
    type: object
  github_com_r0lm0_go-saas-api_internal_agent_model.MessageRole:
    enum:
    - user
    - assistant
    - system
    - tool
    type: string
    x-enum-varnames:
    - MessageRoleUser
    - MessageRoleAssistant
    - MessageRoleSystem
    - MessageRoleTool
  github_com_r0lm0_go-saas-api_internal_agent_model.ToolCall:
    properties:
      arguments:
        additionalProperties: {}
        type: object
      id:
        type: string
      name:
        type: string
    type: object
  github_com_r0lm0_go-saas-api_internal_apikey.APIKey:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      revoked_at:
        type: string
      scopes:
        items:
          type: string
        type: array
      tenant_id:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_r0lm0_go-saas-api_internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_r0lm0_go-saas-api_internal_auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_r0lm0_go-saas-api_internal_auth.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  github_com_r0lm0_go-saas-api_internal_auth.TokenPair:
    properties:
      access_token:
        type: string
      expires_in:
        description: seconds
        type: integer
      refresh_token:
        type: string
    type: object
  github_com_r0lm0_go-saas-api_internal_auth.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      oauth_provider:
        type: string
      oauth_subject:
        type: string
      role:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_r0lm0_go-saas-api_internal_fileupload.Upload:
    properties:
      content_type:
        type: string
      conversation_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      original_name:
        type: string
      size_bytes:
        type: integer
      tenant_id:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/api-keys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.apiKeysResp'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      parameters:
      - description: API key name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api-gateway.apiKeyCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd_api-gateway.apiKeyCreateResp'
      security:
      - BearerAuth: []
      summary: Create a new API key
      tags:
      - api-keys
...
  /api/v1/artifacts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Artifact data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api-gateway.artifactReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_agent_model.Artifact'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new artifact
      tags:
      - agent
...
  /api/v1/auth/callback:
    post:
      consumes:
      - application/json
      parameters:
      - description: OAuth code and state
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.loginResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unified OAuth callback (r3-chat)
      tags:
      - auth
...
  /api/v1/auth/github/callback:
    get:
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.loginResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: GitHub OAuth callback
      tags:
      - auth
...
  /api/v1/auth/google/callback:
    get:
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.loginResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Google OAuth callback
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.loginResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with email and password
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.logoutResp'
      summary: Logout and invalidate refresh token
      tags:
      - auth
  /api/v1/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.meResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - auth
  /api/v1/auth/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.meResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile (r3-chat alias)
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.refreshResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd_api-gateway.registerResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
  /api/v1/chat:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Chat title and model
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new chat (r3-chat)
      tags:
      - agent
...
  /api/v1/conversations:
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.conversationsResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List user conversations
      tags:
      - agent
...
  /api/v1/files:
    get:
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.filesResp'
      security:
      - BearerAuth: []
      summary: List uploaded files
      tags:
      - files
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_fileupload.Upload'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload a file
      tags:
      - files
...
  /api/v1/subscriptions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: User ID
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Plan
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create subscription (r3-chat)
      tags:
      - billing
...
  /api/v1/users/profile:
    put:
      consumes:
      - application/json
      parameters:
      - description: Profile updates
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.meResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
swagger: "2.0"